requirements.txt
openai>=1.2.0
chromadb
streamlit
numpy
pandas
tiktoken
ruff
python-dotenv